Gfx Debugger (#2)
* BitConverter: fix type correcteness in loop

* gbi: get rid of LoadUcode macros

* Gui: add ability to make a GuiImage out of a LUS::Texture

* Add GfxDebugger

* GfxDebugger: add C Bridge

* gfx_pc: expose internal state + intergrate debugger

* Add GfxDebuggerWindow

* Context: integrate GfxDebuggerWindow

* update CMakeLists.txt

* GfxDebugger: move inline code from header to cpp file

+++ extern/ZAPDUtils/Utils/BitConverter.h
@@ -193,12 +193,12 @@ public:
 
 		switch (firstByte) {
 			case 0x37: // v64
-				for (int32_t pos = 0; pos < (romSize / 2); pos++) {
+				for (size_t pos = 0; pos < (romSize / 2); pos++) {
 					((uint16_t*)rom)[pos] = ToUInt16BE(rom, pos * 2);
 				}
 				break;
 			case 0x40: // n64
-				for (int32_t pos = 0; pos < (romSize / 4); pos++) {
+				for (size_t pos = 0; pos < (romSize / 4); pos++) {
 					((uint32_t*)rom)[pos] = ToUInt32BE(rom, pos * 4);
 				}
 				break;
